feature/Types refactoring
