+++ lib/blocs/pokemon_types_bloc.dart
@@ -2,7 +2,7 @@
 import 'dart:async';
 
 import 'package:pogo_diary/blocs/bloc_provider.dart';
-import 'package:pogo_diary/data/pokemon_type_value.dart';
+import 'package:pogo_diary/data/pokemon_types.dart';
 
 class PokemonTypesBloc implements BlocBase {
   final _types = PokemonTypes.values;
