+++ lib/data/pokemon_type_value.dart
@@ -1,20 +1,23 @@
-enum PokemonTypes {
-  Normal,
-  Fighting, 
-  Flying,
-  Poison,
-  Ground,
-  Rock,
-  Bug,
-  Ghost,
-  Steel,
-  Fire,
-  Water,
-  Grass,
-  Electric,
-  Psychic,
-  Ice, 
-  Dragon,
-  Dark,
-  Fairy
+
+class PokemonTypesDataSource {
+  final pokemonTypeValues = const <List<double>>[
+    const <double>[1, 1, 1, 1, 1, 0.8, 1, 0.8, 0.8, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+    const <double>[1.25, 1, 0.8, 0.8, 1, 1.25, 0.8, 0.8, 1,25, 1, 1, 1, 1, 0.8, 1.25, 1, 0.8],
+    const <double>[1, 1.25, 1, 1, 1, 0.8, 1.25, 1, 0.8, 1, 1, 1.25, 0.8, 1, 1, 1, 1, 1],
+    const <double>[1, 1, 1, 0.8, 0.8, 0.8, 1, 0.8, 0.8, 1, 1, 1.25, 1, 1, 1, 1, 1, 1.25],
+    const <double>[1, 1, 0.8, 1.25, 1, 1.25, 0.8, 1, 1.25, 1.25, 1, 0.8, 1.25, 1, 1, 1, 1, 1],
+    const <double>[1, 0.8, 1.25, 1, 0.8, 1, 1.25, 1, 0.8, 1.25, 1, 1, 1, 1, 1.25, 1, 1, 1],
+    const <double>[1, 0.8, 0.8, 0.8, 1, 1, 1, 0.8, 0.8, 0.8, 1, 1.25, 1, 1.25, 1, 1, 1.25, 0.8],
+    const <double>[0.8, 1, 1, 1, 1, 1, 1, 1.25, 1, 1, 1, 1, 1, 1.25, 1, 1, 0.8, 1],
+    const <double>[1, 1, 1, 1, 1, 1.25, 1, 1, 0.8, 0.8, 0.8, 1, 0.8, 1, 1.25, 1, 1, 1.25],
+    const <double>[1, 1, 1, 1, 1, 0.8, 1.25, 1, 1.25, 0.8, 0.8, 1.25, 1, 1, 1.25, 0.8, 1, 1],
+    const <double>[1, 1, 1, 1, 1.25, 1.25, 1, 1, 1, 1.25, 0.8, 0.8, 1, 1, 1, 0.8, 1, 1],
+    const <double>[1, 1, 0.8, 0.8, 1.25, 1.25, 0.8, 1,  0.8, 0.8, 1.25, 0.8, 1, 1, 1, 0.8, 1, 1],
+    const <double>[1, 1, 1.25, 1, 0.8, 1, 1, 1, 1, 1, 1.25, 0.8, 0.8, 1, 1,  0.8, 1, 1],
+    const <double>[1, 1.25, 1, 1.25, 1, 1, 1, 1, 0.8, 1, 1, 1, 1, 0.8, 1, 1, 0.8, 1],
+    const <double>[1, 1, 1.25, 1, 1.25, 1, 1, 1, 0.8, 0.8, 0.8, 1.25, 1, 1, 0.8, 1.25, 1, 1],
+    const <double>[1, 1, 1, 1, 1, 1, 1, 1, 0.8, 0.8, 0.8, 1.25, 1, 1, 0.8, 1.25, 1, 1],
+    const <double>[1, 0.8, 1, 1, 1, 1, 1, 1.25, 1, 1, 1, 1, 1, 1.25, 1, 1, 0.8, 0.8],
+    const <double>[1, 1.25, 1, 0.8, 1, 1, 1, 1, 0.8, 0.8, 1, 1, 1, 1, 1, 1.25, 1.25, 1],
+  ];
 }
