+++ lib/pages/home.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pogo_diary/blocs/bloc_provider.dart';
 import 'package:pogo_diary/blocs/pokemon_types_bloc.dart';
-import 'package:pogo_diary/data/pokemon_type_value.dart';
+import 'package:pogo_diary/data/pokemon_types.dart';
 
 class HomePage extends StatelessWidget {
   @override
